\name{RcmdrqccPlugin-internal}
\title{Internal RcmdrqccPlugin objects}
\alias{checkpkg}
\alias{thereispkg}
\alias{pareto}
\alias{cgXbar}
\alias{cgXbarCCO}
\alias{cgXbarewma}
\alias{cgXbarcusum}
\alias{cgXbarcapabi}
\alias{cgS}
\alias{cgSewma}
\alias{cgScusum}
\alias{cgR}
\alias{cgRewma}
\alias{cgRcusum}
\alias{cgXbarone}
\alias{cgXbaroneewma}
\alias{cgXbaronecusum}
\alias{cgXbaronecapabi}
\alias{cgP}
\alias{cgPCCO}
\alias{cgPewma}
\alias{cgPcusum}
\alias{cgnP}
\alias{cgnPCCO}
\alias{cgnPewma}
\alias{cgnPcusum}
\alias{cgC}
\alias{cgCCCO}
\alias{cgCewma}
\alias{cgCcusum}
\alias{cgU}
\alias{cgUCCO}
\alias{cgUewma}
\alias{cgUcusum}
\alias{cgG}
\alias{cgGewma}
\alias{cgGcusum}


\description{Internal RcmdrqccPlugin objects.}
\details{These are not to be called by the user.}
\keyword{internal}
