\name{RcmdrPlugin.qcc-package}
\alias{RcmdrPlugin.qcc-package}
\alias{RcmdrPlugin.qcc}
\docType{package}
\title{
Install the Demos Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" based on the qcc package, and it provides an integration between the user and the tools of SPC.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.qcc\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2009-01-01\cr
License: \tab GPL version 2 or newer\cr
Encoding: \tab latin1\cr
}
}

\author{
Renan Cortes <renanxc@hotmail.com>, contributions of I. Prestes <isaias.prestes@gmail.com> and Suzi Camey <sucamey@gmail.com>

Maintainer: Renan Cortes <renanxc@hotmail.com>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}